backend/requirements.txt
celery==5.5.0rc5
channels==4.2.0
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
django==5.1.5
djangorestframework-simplejwt==5.4.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
psycopg2==2.9.10
redis==6.0.0b1
requests==2.32.3
